Remove permission update from CustomerService and use old logic in CustomerPermissionManagementService
